responsive completed

+++ src/pages/ReviewCart.jsx
@@ -2,6 +2,8 @@ import { Avatar } from "flowbite-react";
 import React, { useState } from "react";
 import { toast } from "react-toastify";
 import UpdateReview from "./UpdateReview";
+import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
+import { faTrash, faEdit } from "@fortawesome/free-solid-svg-icons";
 
 const ReviewCart = ({ review, reviews, setReviews }) => {
   const [show, setShow] = useState(false);
@@ -11,26 +13,30 @@ const ReviewCart = ({ review, reviews, setReviews }) => {
   };
 
   const handleReviewDelete = (id) => {
-    fetch(`http://localhost:5000/reviews/${id}`, {
-      method: "DELETE",
-    })
-      .then((res) => res.json())
-      .then((data) => {
-        if (data.success) {
-          const remaining = reviews.filter((rev) => rev._id !== id);
-          setReviews(remaining);
-          toast.success("Successfully delete review");
-        }
-      });
+    const confirm = window.confirm("Are you sure want to delete this review");
+
+    if (confirm) {
+      fetch(`http://localhost:5000/reviews/${id}`, {
+        method: "DELETE",
+      })
+        .then((res) => res.json())
+        .then((data) => {
+          if (data.success) {
+            const remaining = reviews.filter((rev) => rev._id !== id);
+            setReviews(remaining);
+            toast.success("Successfully delete review");
+          }
+        });
+    }
   };
 
   return (
     <div
       key={review._id}
-      className="mb-8 p-8 bg-gray-100 border rounded-lg relative"
+      className="mb-8 p-5 md:p-8 bg-gray-100 border rounded-lg relative"
     >
       <div className="flex justify-between items-center">
-        <div className="user-information flex">
+        <div className="user-information md:flex">
           <div className="user-img w-14 mr-4 items-center">
             {review?.img ? (
               <img
@@ -48,19 +54,23 @@ const ReviewCart = ({ review, reviews, setReviews }) => {
           </div>
         </div>
         <div className="update-review-section">
-          <p className="mb-2 text-end">Time: {review.date}</p>
-          <button
-            onClick={handleShowUpdateReview}
-            className="px-5 py-1 text-sm uppercase bg-green-400 text-white rounded-full mr-2"
-          >
-            Edit
-          </button>
-          <button
-            onClick={() => handleReviewDelete(review._id)}
-            className="px-5 py-1 text-sm uppercase bg-red-400 text-white rounded-full"
-          >
-            Delete
-          </button>
+          <p className="mb-2 text-center text-xs md:text-base md:text-end">
+            Time: {review.date}
+          </p>
+          <div className="flex justify-end md:block md:text-end">
+            <FontAwesomeIcon
+              onClick={handleShowUpdateReview}
+              className="text-base bg-gray-300 px-2 py-1 rounded-md"
+              icon={faEdit}
+              title="Edit"
+            />
+            <FontAwesomeIcon
+              onClick={() => handleReviewDelete(review._id)}
+              className="text-base bg-gray-300 px-2 py-1 rounded-md ml-3"
+              icon={faTrash}
+              title="Delete"
+            />
+          </div>
         </div>
       </div>
       <div className="h-[1px] bg-slate-400 mt-5 mb-3"></div>
